Stitcherクラス内のtsposファイルの読み込み処理で、列のデータ型を整数から浮動小数点数に変更しました。これにより、計算精度が向上します。

+++ trainscanner/stitch.py
@@ -219,7 +219,7 @@ class Stitcher:
         tspos = open(self.tsposfile)
         for line in tspos.readlines():
             if len(line) > 0 and line[0] != "@":
-                cols = [int(x) for x in line.split()]
+                cols = [float(x) for x in line.split()]
                 if len(cols) > 0:
                     absx += cols[1]
                     absy += cols[2]
